Move directory for images in datastore to ./images
We need to differentiate between volumes and images so we can enumerate
them independently from eachother.  Moving them from <imageStore(which
is the vch uuid)>/<image id> to <imagestore>/images/<image id>.
